# net/url: use conversion table in shouldEscape