--- a/src/net/url/escape.go
+++ b/src/net/url/escape.go
@@ -11,7 +11,6 @@ package url
 //
 // Please be informed that for now shouldEscape does not check all
 // reserved characters correctly. See golang.org/issue/5684.
-//
 // See src/net/url/escape_gen.go for more info information.
 func shouldEscape(c byte, mode encoding) bool {
 	if shouldEscapeTable[int(mode*256)+int(c)] == 0 {