Use conversion table to speed up shouldEscape.

benchstat:
name                 old time/op    new time/op    delta
String-8               3.45µs ± 0%    2.42µs ± 0%  -29.94%  (p=0.000 n=8+9)
QueryEscape/#00-8      41.6ns ± 1%    28.8ns ± 0%  -30.74%  (p=0.000 n=10+9)
QueryEscape/#01-8       275ns ± 0%      98ns ± 0%  -64.59%  (p=0.000 n=9+10)
QueryEscape/#02-8       157ns ± 1%      71ns ± 1%  -54.88%  (p=0.000 n=9+8)
QueryEscape/#03-8       355ns ± 1%     138ns ± 0%  -61.24%  (p=0.000 n=9+10)
QueryEscape/#04-8      2.24µs ± 1%    0.84µs ± 0%  -62.22%  (p=0.000 n=10+10)
PathEscape/#00-8       65.1ns ± 1%    35.9ns ± 0%  -44.87%  (p=0.000 n=10+9)
PathEscape/#01-8        283ns ± 1%     103ns ± 1%  -63.68%  (p=0.000 n=10+10)
PathEscape/#02-8        158ns ± 0%      70ns ± 0%  -55.72%  (p=0.000 n=7+10)
PathEscape/#03-8        350ns ± 0%     137ns ± 1%  -60.76%  (p=0.000 n=10+10)
PathEscape/#04-8       2.27µs ± 1%    0.85µs ± 0%  -62.63%  (p=0.000 n=10+9)
QueryUnescape/#00-8    41.8ns ± 0%    39.6ns ± 0%   -5.13%  (p=0.000 n=9+9)
QueryUnescape/#01-8     140ns ± 1%     146ns ± 1%   +4.06%  (p=0.000 n=10+10)
QueryUnescape/#02-8     109ns ± 1%     101ns ± 0%   -7.00%  (p=0.000 n=10+7)
QueryUnescape/#03-8     244ns ± 1%     228ns ± 1%   -6.85%  (p=0.000 n=9+10)
QueryUnescape/#04-8    1.36µs ± 1%    1.29µs ± 1%   -5.36%  (p=0.000 n=10+10)
PathUnescape/#00-8     42.6ns ± 1%    41.7ns ± 1%   -2.11%  (p=0.000 n=9+10)
PathUnescape/#01-8      142ns ± 1%     147ns ± 0%   +3.23%  (p=0.000 n=10+6)
PathUnescape/#02-8      108ns ± 0%     101ns ± 0%   -6.48%  (p=0.000 n=9+6)
PathUnescape/#03-8      244ns ± 1%     228ns ± 0%   -6.71%  (p=0.000 n=10+10)
PathUnescape/#04-8     1.30µs ± 1%    1.26µs ± 1%   -2.52%  (p=0.000 n=10+10)
Split-8                7.34ns ± 1%    7.40ns ± 1%   +0.88%  (p=0.000 n=9+8)